+++ uui/src/lib/shared/theme/breakpoints/default-breakpoints.ts
@@ -5,5 +5,5 @@ export const UUI_BREAKPOINTS: Breakpoints = {
   $sm: '600px',
   $md: '900px',
   $lg: '1200px',
-  $xl: '1200px'
+  $xl: '1536px'
 };